Update SparkFun_LED_8x7.h
Adds support for ATmega32U4.

I purchased the wrong board for the 8x7 LED Array from SparkFun, and instead of buying the correct board, I was determined to figure out how to make this work.

It's been about 1.5 years since I edited this on my laptop. I'm recording it before I lose it!